Fix sample click on end
Previously when a sample ended, the sampler would instantly reset that sampler to the start of the sample. If the sample didn't start at 0, it would cause an audible click. Now the sampler waits at the end of the sample, and doesn't jump to the start until the sample is triggered again.